requirements.txt
discord.py[voice]
discord-py-slash-command
python-dotenv
GitPython